Refactored and cleaned up Navbar
