--- src/navbar/LiveBanner/queries.js
@@ -1,10 +1,0 @@
-import gql from 'graphql-tag'
-
-export const GET_LIVE_STREAM = gql`
-  query getLiveStream {
-    liveStream {
-      isLive
-      eventStartTime
-    }
-  }
-`
